Push to prod (#502)
* Bump elliptic from 6.5.2 to 6.5.3 in /backend

Bumps [elliptic](https://github.com/indutny/elliptic) from 6.5.2 to 6.5.3.
- [Release notes](https://github.com/indutny/elliptic/releases)
- [Commits](https://github.com/indutny/elliptic/compare/v6.5.2...v6.5.3)


* Bump elliptic from 6.5.2 to 6.5.3 in /frontend

Bumps [elliptic](https://github.com/indutny/elliptic) from 6.5.2 to 6.5.3.
- [Release notes](https://github.com/indutny/elliptic/releases)
- [Commits](https://github.com/indutny/elliptic/compare/v6.5.2...v6.5.3)


* fixed applications not being updated properly

* lint fix

* changes as per the PR review

* fixes #350; testnet chain id

* Fixing merge issue

* Completed Network Status tests

* Added pocket CLI confirmation test coverage

* resolved typos

* Bump bcrypt from 4.0.1 to 5.0.0 in /backend

Bumps [bcrypt](https://github.com/kelektiv/node.bcrypt.js) from 4.0.1 to 5.0.0.
- [Release notes](https://github.com/kelektiv/node.bcrypt.js/releases)
- [Changelog](https://github.com/kelektiv/node.bcrypt.js/blob/master/CHANGELOG.md)
- [Commits](https://github.com/kelektiv/node.bcrypt.js/compare/v4.0.1...v5.0.0)


* Completed Nodes tests

* Added blockTime verification

* Added BE support to the tests, cases for edit and remove

* Added another remove scenario

* Added base file and structure for APPs sectio, and node section for CLI verification

* Completed APPs section tests

* Fixing the nodes tests as per Vale's suggestion

* Fixed Valeria's comments

* #359; gateway settings feature

* fixes #347; misspelling in gateway settings whitelists

* Bump bl from 2.2.0 to 2.2.1 in /backend

Bumps [bl](https://github.com/rvagg/bl) from 2.2.0 to 2.2.1.
- [Release notes](https://github.com/rvagg/bl/releases)
- [Commits](https://github.com/rvagg/bl/compare/v2.2.0...v2.2.1)


* Added jwt expiration and refresh token expiration env variables, added auth header check for each request

* Added access and refresh token generation when signing in, signing in response changed to include session tokens, added decode token error handlers

* added serviceHelper with an axios interceptor to include the Authorization headers with both access and refresh tokens

* added access and refresh tokens to the user cache flow, updated login save user in cache

* Added express jwt

* added proper generate and renew process for the tokens, added frontend helper for services to manage the tokens

* added jwt interceptor on the backend, added proper flows for the jwt

* added saveUserSessionInCache for FE, updated axios inverceptor for request, added one for responses, fixed session renew function, fixed backend interceptor for requests

* Adding support for custom headers

* Added exclusion api paths for the jwt session checks

* Added security checks for the app, node, payment and user api/services

* updated service helper to include session cached user email and save the new session when is updated from the backend

* removed some endpoints from the session exclusion list, fixed app and node details page failing to load

* fixed app not being properly updated

* fixes #363, #376; gateway settings page fixes

* class to className

* Closing issues

Close #367
Close #370
Close #369
Close #367

* Fixing lint issues

* fixes #346; prevents selecting same question more than once

* #381; linting

* #346; changing method of disctinctness

* #346; lint

* __updateTransaction now marks the transaction as completed

* fixed #383, fixed #384, updated pocket-js

* post transactions now are marked as completed, fixed wrong email template ids for some actions, added missing payment template ids

* added missing jwt env variables to testnet and prod yml

* fixes #246; changing color and text of unjail notice

* fixes #390; security alerts

* Fixes #393; gateway settings not saving

* Adding verification when trying to unstake a node

Close #392

* Adding missing code

* Adding extra verification

* Fixing import duplicated nodes

Close #401

* Fixing Lint

* Removing console and verifying if app exists before import

* Removing unused var

* Fixed no email being sent to new created apps

* added content to the welcome back message

* added link to the pocket logo to the main site

* fixed #411

* Removing duplicated email

Close #398

* fixed #420

* Fixing issues

Close #424
Close #419

* Fixing lint

* Fixing issue for deleting Payment Method

Close #424

* added husky to run es lint before committing any change, removed npm package-lock.json

* Fixing UI issue

Close #404

* #403; adding session timeout to frontend

* fixed #421 #435

* fixed lint issues

* Fixing issue paying with just tokens

* Fixing lint

* closes #409

* disabled providers call in the frontend, added try/catch to saveUserInCache

* Fixes #412; wrong link for token agreement

* Fixing payment search issue

Close #423

* Fixing lint

* Adding support for sending emails after payment

Close #432

* Fixing lint issues

* #409 done

* added map area to the pocket logo

* fixed lint

* updated min and max relays per day, added step to the app relay slider

* fixed #433

* Fixed user can't create security questions, fixed #453

* fixed #460, staked tokens amount is not more accurate in main list and app details

* Replaced landing page text, removed message from welcome message.

* fixed lint issue

* pr review change

* Fixing updatingStatus flag issue

Close #451

* Fixing lint issue

* multiple UI changes/fixes for checkout and invoices

* updated minimum and maximum RPD, updated select relays aprox PRD/POKT cost

* fixed getApplication error not being handled, fixed wrong message and flow while importing an app

* fixed lint

* fixed purchase app/node data format

* added check for login if the user is validated

* Fixing unjail issue

Close #473

* Fixing lint issues

* Fixing unjail issue

Close #473

* Fixing lint issues

* Fixing import

* Fixing account not found issues

* Fixing lint

Close #401
Close #472
Close  #461

* Fixing lint warning

* fixed wrong amount of tokens after purchase in the invoice

* fixed app unstake failing

* fixed missing account in pocket instance, fixed wrong calculation while using current balance, ppk is now being saved in the appDetails too, chain list is now displayed if app is staked

* removed convertion for payment service, added pokt currency type, added proper purchased tokens to the invoice second page, fixed app summary format issue

* missing change

* fixed bad integer and round miscalculation

* fixed lints

* typo fix

* changed RPD formula

* fixed new upokt value not being returned in a scenario for the rpd formula

* fixed some typos

* updated free tier stake amount

* Fixing Node issues

Close #489
Close #488

* Fixing lint

* fixed wrong app data formatting while importing an account

* improved use my balance input

* fixes #500; changes prod url

* Adding Provider arch

* Adding Provider arch

* Fixing lint issues

+++ backend/src/apis/ApplicationApi.js
@@ -4,10 +4,12 @@ import {apiAsyncWrapper, getOptionalQueryOption, getQueryOption} from "./_helper
 import EmailService from "../services/EmailService";
 import PaymentService from "../services/PaymentService";
 import ApplicationCheckoutService from "../services/checkout/ApplicationCheckoutService";
-import {CoinDenom} from "@pokt-network/pocket-js";
+import UserService from "../services/UserService";
+import numeral from "numeral";
 
 const router = express.Router();
 
+const userService = new UserService();
 const applicationService = new ApplicationService();
 const applicationCheckoutService = ApplicationCheckoutService.getInstance();
 const paymentService = new PaymentService();
@@ -38,7 +40,7 @@ router.post("/account", apiAsyncWrapper(async (req, res) => {
     link: `${data.applicationBaseLink}/${data.applicationData.address}`
   };
 
-  if (emailAction === "imported") {
+  if (emailAction) {
     await EmailService
       .to(application.contactEmail)
       .sendCreateOrImportAppEmail(emailAction, application.contactEmail, applicationEmailData);
@@ -57,9 +59,13 @@ router.put("/:applicationId", apiAsyncWrapper(async (req, res) => {
   /** @type {{applicationId:string}} */
   const params = req.params;
 
-  const updated = await applicationService.updateApplication(params.applicationId, data);
+  if (await applicationService.verifyApplicationBelongsToClient(params.applicationId, req.headers.authorization)) {
+    const updated = await applicationService.updateApplication(params.applicationId, data);
 
-  res.send(updated);
+    res.send(updated);
+  } else {
+    res.status(400).send("Application doesn't belong to the client account.");
+  }
 }));
 
 /**
@@ -71,20 +77,24 @@ router.post("/:applicationId", apiAsyncWrapper(async (req, res) => {
   /** @type {{user:string, appsLink:string}} */
   const bodyData = req.body;
 
-  const application = await applicationService.deleteApplication(data.applicationId, bodyData.user);
+  if (await applicationService.verifyApplicationBelongsToClient(data.applicationId, req.headers.authorization) ) {
+    const application = await applicationService.deleteApplication(data.applicationId, bodyData.user);
 
-  if (application) {
-    const applicationEmailData = {
-      name: application.name,
-      appsLink: bodyData.appsLink
-    };
+    if (application) {
+      const applicationEmailData = {
+        name: application.name,
+        appsLink: bodyData.appsLink
+      };
 
-    await EmailService
-      .to(application.contactEmail)
-      .sendAppDeletedEmail(application.contactEmail, applicationEmailData);
+      await EmailService
+        .to(application.contactEmail)
+        .sendAppDeletedEmail(application.contactEmail, applicationEmailData);
+    }
+
+    res.send(application !== undefined);
+  } else {
+    res.status(400).send("Application doesn't belong to the client account.");
   }
-
-  res.send(application !== undefined);
 }));
 
 /**
@@ -124,9 +134,14 @@ router.get("/:applicationAccountAddress", apiAsyncWrapper(async (req, res) => {
 router.get("/client/:applicationId", apiAsyncWrapper(async (req, res) => {
   /** @type {{applicationId:string}} */
   const data = req.params;
-  const application = await applicationService.getClientApplication(data.applicationId);
 
-  res.json(application);
+  if (await applicationService.verifyApplicationBelongsToClient(data.applicationId, req.headers.authorization)) {
+    const application = await applicationService.getClientApplication(data.applicationId);
+
+    res.json(application);
+  } else {
+    res.status(400).send("Application doesn't belong to the client account.");
+  }
 }));
 
 /**
@@ -166,9 +181,13 @@ router.post("/user/all", apiAsyncWrapper(async (req, res) => {
   /** @type {{user: string}} */
   const data = req.body;
 
-  const applications = await applicationService.getUserApplications(data.user, limit, offset);
+  if (await userService.verifySessionForClient(req.headers.authorization, data.user)) {
+    const applications = await applicationService.getUserApplications(data.user, limit, offset);
 
-  res.json(applications);
+    res.json(applications);
+  } else {
+    res.status(400).send("Application list doesn't belong to the provided client account.");
+  }
 }));
 
 /**
@@ -194,22 +213,22 @@ router.post("/freetier/stake", apiAsyncWrapper(async (req, res) => {
 /**
  * Unstake a free tier application.
  */
-router.post("/freetier/unstake", apiAsyncWrapper(async (req, res) => {
-  /** @type {{unstakeInformation: {application_id: string}, applicationLink: string}} */
-  const data = req.body;
+// router.post("/freetier/unstake", apiAsyncWrapper(async (req, res) => {
+//   /** @type {{unstakeInformation: {application_id: string}, applicationLink: string}} */
+//   const data = req.body;
 
-  const unstakeInformation = data.unstakeInformation;
+//   const unstakeInformation = data.unstakeInformation;
 
-  await applicationService.unstakeFreeTierApplication(unstakeInformation, data.applicationLink);
+//   await applicationService.unstakeFreeTierApplication(unstakeInformation, data.applicationLink);
 
-  res.send(true);
-}));
+//   res.send(true);
+// }));
 
 /**
  * Stake an application.
  */
 router.post("/custom/stake", apiAsyncWrapper(async (req, res) => {
-  /** @type {{applicationId: string, appStakeTransaction: {address: string, raw_hex_bytes: string}, paymentId: string, applicationLink: string, gatewayAATSignature: string}} */
+  /** @type {{applicationId: string, appStakeTransaction: {address: string, raw_hex_bytes: string}, paymentId: string, applicationLink: string, gatewayAATSignature: string, upoktToStake: string}} */
   const data = req.body;
   const paymentHistory = await paymentService.getPaymentFromHistory(data.paymentId);
 
@@ -220,25 +239,28 @@ router.post("/custom/stake", apiAsyncWrapper(async (req, res) => {
     const appStakeTransaction = data.appStakeTransaction;
     const application = await applicationService.getClientApplication(data.applicationId);
 
-    const item = paymentHistory.getItem();
-    const amountToSpent = applicationCheckoutService.getMoneyToSpent(parseInt(item.maxRelays));
-    const poktStaked = applicationCheckoutService.getPoktToStake(amountToSpent, CoinDenom.Pokt).toString();
-    const uPoktStaked = applicationCheckoutService.getPoktToStake(amountToSpent, CoinDenom.Upokt).toString();
+    if (await applicationService.verifyApplicationBelongsToClient(data.applicationId, req.headers.authorization)) {
+      const item = paymentHistory.getItem();
+      // For the email, convert to pokt
+      const poktStaked = data.upoktToStake / 1000000;
 
-    const applicationEmailData = {
-      name: application.pocketApplication.name,
-      link: data.applicationLink
-    };
+      const applicationEmailData = {
+        name: application.pocketApplication.name,
+        link: data.applicationLink
+      };
 
-    const paymentEmailData = {
-      amountPaid: paymentHistory.amount,
-      maxRelayPerDayAmount: item.maxRelays,
-      poktStaked: poktStaked
-    };
+      const paymentEmailData = {
+        amountPaid: numeral(paymentHistory.amount / 100).format("0,0.00"),
+        maxRelayPerDayAmount: numeral(item.maxRelays).format("0,0.00"),
+        poktStaked: numeral(poktStaked).format("0,0.000000")
+      };
 
-    await applicationService.stakeApplication(appStakeTransaction.address, uPoktStaked, appStakeTransaction, application, applicationEmailData, paymentEmailData, data.gatewayAATSignature);
+      await applicationService.stakeApplication(appStakeTransaction.address, data.upoktToStake, appStakeTransaction, application, applicationEmailData, paymentEmailData, data.gatewayAATSignature);
 
-    res.send(true);
+      res.send(true);
+    } else {
+      res.status(400).send("Application doesn't belong to the provided client account.");
+    }
   } else {
     // Return error if payment was unsuccessful
     throw new Error("Error processing payment, please try a different method");
@@ -257,7 +279,7 @@ router.post("/custom/unstake", apiAsyncWrapper(async (req, res) => {
   } = data;
 
   // Submit Unstake transaction to the pocket network
-  await applicationService.unstakeApplication(appUnstakeTransaction, applicationLink);
+  await applicationService.unstakeApplication(appUnstakeTransaction, applicationLink, req.headers.authorization);
 
   // Respond
   res.send(true);
@@ -270,9 +292,13 @@ router.get("/freetier/aat/:applicationId", apiAsyncWrapper(async (req, res) => {
   /** @type {{applicationId:string}} */
   const data = req.params;
 
-  const aat = await applicationService.getFreeTierAAT(data.applicationId);
+  if (await applicationService.verifyApplicationBelongsToClient(data.applicationId, req.headers.authorization)) {
+    const aat = await applicationService.getFreeTierAAT(data.applicationId);
 
-  res.json(aat);
+    res.json(aat);
+  } else {
+    res.status(400).send("Application doesn't belong to the provided client account.");
+  }
 }));
 
 
@@ -280,9 +306,13 @@ router.post("/update/gateway/settings", apiAsyncWrapper(async (req, res) => {
   const data = req.body;
   const applicationId = data.id;
 
-  const updated = await applicationService.updateApplication(applicationId, data);
+  if (await applicationService.verifyApplicationBelongsToClient(applicationId, req.headers.authorization)) {
+    const updated = await applicationService.updateApplication(applicationId, data);
 
-  res.send(updated);
+    res.send(updated);
+  } else {
+    res.status(400).send("Application doesn't belong to the provided client account.");
+  }
 }));
 
 export default router;
