+++ backend/src/services/PaymentService.js
@@ -1,5 +1,5 @@
 import BaseService from "./BaseService";
-import {get_default_payment_provider} from "../providers/payment/Index";
+import {get_default_payment_provider, getTokenPaymentProvider, providerType} from "../providers/payment/Index";
 import {CardPaymentMethod, Payment, PaymentCurrencies, PaymentResult} from "../providers/payment/BasePaymentProvider";
 import {BillingDetails, PaymentHistory, PaymentMethod} from "../models/Payment";
 import UserService from "./UserService";
@@ -14,7 +14,6 @@ export default class PaymentService extends BaseService {
   constructor() {
     super();
 
-    this._paymentProvider = get_default_payment_provider();
     this.userService = new UserService();
   }
 
@@ -27,15 +26,28 @@ export default class PaymentService extends BaseService {
    * @param {*} item Item to pay.
    * @param {number} amount Amount intended to be collected by this payment.
    * @param {string} to For what is the payment (Apps or Node).
+   * @param {number} tokens Tokens used for this payment.
    *
    * @returns {Promise<PaymentResult>} A payment result of intent.
    * @private
    * @async
    */
-  async __createPocketPaymentIntent(userCustomerID, type, currency, item, amount, to) {
+  async __createPocketPaymentIntent(userCustomerID, type, currency, item, amount, to, tokens) {
     const description = `Acquiring ${to.toLowerCase() === "application" ? "Max Relays Per Day" : "Validator Power"} for ${to}`;
 
-    return this._paymentProvider.createPaymentIntent(userCustomerID, type, currency, item, amount, description);
+    if (amount === 0) {
+      return this.__getPaymentProvider(providerType.token).createPaymentIntent(userCustomerID, type, currency, item, amount, description, tokens);
+    } else {
+      return this.__getPaymentProvider(providerType.stripe).createPaymentIntent(userCustomerID, type, currency, item, amount, description, tokens);
+    }
+  }
+
+  __getPaymentProvider(type) {
+    if (type === providerType.token) {
+      return getTokenPaymentProvider();
+    } else {
+      return get_default_payment_provider();
+    }
   }
 
   /**
@@ -47,33 +59,37 @@ export default class PaymentService extends BaseService {
    * @param {*} item Item to pay.
    * @param {number} amount Amount intended to be collected by this payment.
    * @param {string} itemType Item type for payment.
+   * @param {number} tokens Tokens used for this payment.
    *
    * @returns {Promise<PaymentResult | boolean>} A payment result of intent.
    * @throws {DashboardValidationError} if validation fails.
    * @async
    */
-  async __createPocketPaymentForItem(userEmail, type, currency, item, amount, itemType) {
+  async __createPocketPaymentForItem(userEmail, type, currency, item, amount, itemType, tokens) {
+
     if (!Payment.validate({type, currency, item, amount})) {
       return false;
     }
+
+    const provType = amount === 0 ? providerType.token : providerType.stripe;
 
     // Getting user customer from user, a customer is required by stripe.
     let userCustomerID = await this.userService.getUserCustomerID(userEmail);
 
     if (!userCustomerID) {
-      const userCustomer = await this._paymentProvider.createCustomer(userEmail);
+      const userCustomer = await this.__getPaymentProvider(provType).createCustomer(userEmail);
 
       userCustomerID = userCustomer.id;
       await this.userService.saveCustomerID(userEmail, userCustomerID);
     }
 
-    const amountFixed = Math.round(amount);
+    const amountFixed = Math.round(amount.toFixed(2)*100);
     const paymentItem = {
       ...item,
       type: itemType
     };
 
-    return this.__createPocketPaymentIntent(userCustomerID, type, currency, paymentItem, amountFixed, itemType);
+    return this.__createPocketPaymentIntent(userCustomerID, type, currency, paymentItem, amountFixed, itemType, tokens);
   }
 
   /**
@@ -155,18 +171,35 @@ export default class PaymentService extends BaseService {
   /**
    * Delete a payment method from DB.
    *
-   * @param {string} paymentMethodID Payment method ID
+   * @param {string} paymentMethodID Payment method ID.
+   * @param {string} authHeader Authorization header.
    *
    * @returns {Promise<boolean>} If was deleted or not.
    * @async
    */
-  async deletePaymentMethod(paymentMethodID) {
-    const filter = {id: paymentMethodID};
+  async deletePaymentMethod(paymentMethodID, authHeader) {
+    const filter = {"paymentMethod.id": paymentMethodID};
+    const userEmail = authHeader.split(", ")[2].split(" ")[1];
+    let belongsToClient = false;
 
-    /** @type {{result: {n:number, ok: number}}} */
-    const result = await this.persistenceService.deleteEntities(PAYMENT_METHOD_COLLECTION_NAME, filter);
+    if (userEmail) {
+      const paymentMethods = await this.getUserPaymentMethods(userEmail);
 
-    return result.result.ok === 1;
+      paymentMethods.forEach(method => {
+
+        if (method.id.toString() === paymentMethodID.toString()) {
+          belongsToClient = true;
+        }
+      });
+
+      if (belongsToClient) {
+        /** @type {{result: {n:number, ok: number}}} */
+        const result = await this.persistenceService.deleteEntities(PAYMENT_METHOD_COLLECTION_NAME, filter);
+
+        return result.result.ok === 1;
+      }
+    }
+    return belongsToClient;
   }
 
   /**
@@ -183,9 +216,9 @@ export default class PaymentService extends BaseService {
    * @async
    */
   async createPocketPaymentIntentForApps(paymentIntentData) {
-    const {user, type, currency, item, amount} = paymentIntentData;
+    const {user, type, currency, item, amount, tokens} = paymentIntentData;
 
-    return this.__createPocketPaymentForItem(user, type, currency, item, amount, "Application");
+    return this.__createPocketPaymentForItem(user, type, currency, item, amount, "Application", tokens);
   }
 
   /**
@@ -202,9 +235,9 @@ export default class PaymentService extends BaseService {
    * @async
    */
   async createPocketPaymentIntentForNodes(paymentIntentData) {
-    const {user, type, currency, item, amount} = paymentIntentData;
+    const {user, type, currency, item, amount, tokens} = paymentIntentData;
 
-    return this.__createPocketPaymentForItem(user, type, currency, item, amount, "Node");
+    return this.__createPocketPaymentForItem(user, type, currency, item, amount, "Node", tokens);
   }
 
   /**
@@ -215,10 +248,11 @@ export default class PaymentService extends BaseService {
    * @param {number} [offset] Offset of query.
    * @param {string} [fromDate] From created date.
    * @param {string} [toDate] To created date.
+   * @param {string} [paymentID] Payment ID.
    *
    * @returns {Promise<PaymentHistory[]>} List of Payment history.
    */
-  async getPaymentHistory(user, limit, offset = 0, fromDate = "", toDate = "") {
+  async getPaymentHistory(user, limit, offset = 0, fromDate = "", toDate = "", paymentID = "") {
     let filter = {user};
     let dateFilter = {};
 
@@ -234,9 +268,43 @@ export default class PaymentService extends BaseService {
       filter["createdDate"] = dateFilter;
     }
 
+    if (paymentID) {
+      filter["paymentID"] = paymentID;
+    }
 
     return (await this.persistenceService.getEntities(PAYMENT_HISTORY_COLLECTION_NAME, filter, limit, offset))
       .map(PaymentHistory.createPaymentHistory);
+  }
+
+  /**
+   * Verify if the payment belongs to the client's account using an payment id
+   *
+   * @param {string} paymentId Application Identifier.
+   * @param {string} authHeader Authorization header.
+   *
+   * @returns {Promise<boolean>} True if the payment belongs to the client account or false otherwise.
+   * @async
+   */
+  async verifyPaymentBelongsToClient(paymentId, authHeader) {
+    // Retrieve the session tokens from the auth headers
+    const accessToken = authHeader.split(", ")[0].split(" ")[1];
+    const userEmail = authHeader.split(", ")[2].split(" ")[1];
+
+    if (accessToken && userEmail) {
+      const payload = await this.userService.decodeToken(accessToken, true);
+
+      if (payload instanceof DashboardValidationError) {
+        throw payload;
+      }
+      // Use token email to retrieve a list of the payments
+      const paymentHistory = await this.getPaymentFromHistory(paymentId);
+
+      if (paymentHistory.user.toString() === userEmail.toString() && userEmail.toString() === payload.email.toString()) {
+        return true;
+      }
+    }
+
+    return false;
   }
 
   /**
@@ -248,12 +316,16 @@ export default class PaymentService extends BaseService {
    * @param {number} amount Amount.
    * @param {*} item Item bought.
    * @param {string} user User that belongs the payment.
+   * @param {number} tokens Tokens used for this payment.
    *
    * @returns {Promise<boolean>} If payment was saved or not.
    * @async
    */
-  async savePaymentHistory(createdDate, paymentID, currency, amount, item, user) {
+  async savePaymentHistory(createdDate, paymentID, currency, amount, item, user, tokens) {
+
     const {pokt_market_price: poktPrice} = Configurations.pocket_network;
+    const token = tokens === undefined ? 0 : tokens;
+
     const paymentHistory = PaymentHistory.createPaymentHistory({
       createdDate,
       paymentID,
@@ -261,7 +333,8 @@ export default class PaymentService extends BaseService {
       amount,
       item,
       user,
-      poktPrice
+      poktPrice,
+      tokens: token
     });
 
     if (await this.paymentHistoryExists(paymentHistory)) {
@@ -274,6 +347,32 @@ export default class PaymentService extends BaseService {
     return result.result.ok === 1;
   }
 
+  /**
+   * Update payment history record with the printable data object.
+   *
+   * @param {string} paymentId Payment Identifier.
+   * @param {string} userEmail User email.
+   * @param {object} printableData Payment invoice printable data.
+   *
+   * @returns {Promise<boolean>} If payment was marked or not.
+   * @async
+   */
+  async updatePaymentWithPrintableData(paymentId, userEmail, printableData) {
+    const filter = {paymentID: paymentId, user: userEmail};
+
+    const paymentHistoryDB = await this.persistenceService.getEntityByFilter(PAYMENT_HISTORY_COLLECTION_NAME, filter);
+
+    if (!paymentHistoryDB) {
+      return false;
+    }
+
+    paymentHistoryDB.printableData = printableData;
+
+    /** @type {{result: {n:number, ok: number}}} */
+    const result = await this.persistenceService.updateEntity(PAYMENT_HISTORY_COLLECTION_NAME, filter, paymentHistoryDB);
+
+    return result.result.ok === 1;
+  }
 
   /**
    * Mark payment as success on history.
@@ -318,7 +417,7 @@ export default class PaymentService extends BaseService {
 
     if (dbPaymentMethods) {
       const paymentMethodIds = dbPaymentMethods.map(_ => _.paymentMethod.id);
-      const paymentMethods = await this._paymentProvider.retrieveCardPaymentMethods(paymentMethodIds);
+      const paymentMethods = await this.__getPaymentProvider(providerType.stripe).retrieveCardPaymentMethods(paymentMethodIds);
 
       return Promise.all(paymentMethods);
     }
@@ -339,8 +438,12 @@ export default class PaymentService extends BaseService {
     const dbPaymentHistory = await this.persistenceService.getEntityByFilter(PAYMENT_HISTORY_COLLECTION_NAME, filter);
 
     if (dbPaymentHistory) {
-      return PaymentHistory.createPaymentHistory(dbPaymentHistory);
-    }
+      const paymentHistory = PaymentHistory.createPaymentHistory(dbPaymentHistory);
+
+      if (paymentHistory) {
+        return paymentHistory;
+      }
+  }
 
     return null;
   }
